:art: Code split notes model into its own file, update require statements
