+++ server/models/notesModel.js
@@ -1,0 +1,12 @@
+const mongoose = require('mongoose');
+// Schema for our notes document
+const noteSchema = new Schema({
+  owner_id: { type: Schema.Types.ObjectId, ref: 'user' },
+  title: String,
+  content: String,
+});
+
+// Create our notes model linking the notesSchema and notes document
+const Notes = mongoose.model('note', noteSchema);
+
+module.exports = Notes
